Rename .py.3only files to .py
